+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - [device]: Added new method addClusterServerFromList
 - [frontend]: Fetch data every 10 minutes
+- [device]: Added ModeSelectClusterServer
 
 
 <a href="https://www.buymeacoffee.com/luligugithub">